feat: add terraform example (#7)
Sorry @DanielMSchmidt, in order to manage correctly some directories in gitignore I need to do a complete re-init of my Branch. 
In any case, here I've added the change you suggested. 
For complete the "provider-agnostic" idea on the main.tf file, I have this idea: We can manage a parameter, for example `terraformProvider` and based on this we can create different `main.tf` files, that correspond to the right one ( a file for azure, a file for aws, etc..).
In this way we can automate it using projen. 
The fact is that if the main.tf file is commented and without anything, I think that the validation command may go in error so in order to launch the `terraform validate && terraform plan` command correctly the file need to be valorized correctly. 

WDYT?