chore!: upgrade to cdktf 0.18.0 (#86)
This PR initiates the upgrade of CDKTF from version `null` to version
`0.18.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update `@cdktf/tf-module-stack` to a version compatible with
`cdktf@0.18.0`
[here](https://github.com/cdktf/projen-cdktf-hybrid-construct/blob/ad6fff04398e309fa5f09e12b1c658f1a51937d5/src/hybrid-module.ts#L194).
Look up the version
[here](https://www.npmjs.com/package/@cdktf/tf-module-stack?activeTab=versions).
- [x] Run `npx projen build`
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Nara Kasbergen <nara.kasbergen@gmail.com>