chore!: upgrade to cdktf 0.21.0 (#219)
This PR initiates the upgrade of CDKTF from version `0.20.0` to version
`0.21.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update `@cdktf/tf-module-stack` to a version compatible with
`cdktf@0.21.0`
[here](https://github.com/cdktf/projen-cdktf-hybrid-construct/blob/62c33dca3eaa249b519edd82fa0889ffcd12c6eb/src/hybrid-module.ts#L194).
Look up the version
[here](https://www.npmjs.com/package/@cdktf/tf-module-stack?activeTab=versions).
- [x] Run `npx projen build`
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>